+++ lib/disposable_cached_images.dart
@@ -2,6 +2,7 @@ library disposable_cached_images;
 
 import 'dart:io';
 import 'dart:typed_data';
+import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -10,8 +11,8 @@ import 'package:http/http.dart' as http;
 import 'package:path_provider/path_provider.dart';
 
 part './disposable_cached_image_provider_abstract.dart';
-part './disposable_image_widgets.dart';
+part './disposable_image_widget.dart';
 part './globals.dart';
 part './image_data_base.dart';
+part './image_provider.dart';
 part './image_provider_state.dart';
-part './image_providers.dart';