+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 ## 0.0.3
 
 Support assets images.
+
 Remove Dio in favor of http.
+
+## 0.0.5
+
+Add maxCacheWidth and maxCacheHeight to resize images
+
+Update documentation and example
+
+- breaking changes:
+
+Integrate the asset image with the network image to use one Widget for both
+
+Improved error handling